+++ Changelog.md
@@ -16,6 +16,13 @@
 
 # Dev
 
+## Steps
+
+* `OpenROAD.CTS`
+  * Added flags `CTS_OBSTRUCTION_AWARE` and `CTS_BALANCE_LEVELS`
+  * Added `CTS_SINK_BUFFER_MAX_CAP_DERATE_PCT`
+  * Added `CTS_DELAY_BUFFER_DERATE_PCT`
+
 ## Misc Enhancements/Bugfixes
 
 * `openlane.state`